Move pocketbase to server and provide and fix 500
PocketBase is now provided from the server side with a plugin (following
PocketBase JS SDK documentation), making it
more secure and practical than the actual implementation with
provide/inject.

A 500 was being thrown due to the way that the NaiveUI components were
imported, it is fixed now. As the objective is to completely move out of
NaiveUI in favor of Equal, NaiveUI's components are now all registered
with the use of a plugin, which will later be deleted.